Add managed credentials mode

+++ src/UI/class-settings-page.php
@@ -205,7 +205,8 @@ final class Settings_Page {
 	 * @since 3.2.0
 	 */
 	private function initialize_basic_section(): void {
-		$section_key = 'basic-section';
+		$are_credentials_managed = Parsely::are_credentials_managed();
+		$section_key             = 'basic-section';
 
 		add_settings_section(
 			$section_key,
@@ -218,11 +219,12 @@ final class Settings_Page {
 		$field_id   = 'apikey';
 		$field_args = array(
 			'option_key'    => $field_id,
-			'help_text'     => __( 'Your Site ID is typically your own site domain without <code>http(s)://</code> prefixes or trailing <code>/</code> (e.g. <code>mydomain.com</code>).', 'wp-parsely' ),
+			'help_text'     => $are_credentials_managed ? '' : __( 'Your Site ID is typically your own site domain without <code>http(s)://</code> prefixes or trailing <code>/</code> (e.g. <code>mydomain.com</code>).', 'wp-parsely' ),
 			'label_for'     => $field_id,
 			'optional_args' => array(
 				'required'    => 'required',
 				'placeholder' => 'mydomain.com',
+				'disabled'    => $are_credentials_managed ? true : false,
 			),
 
 		);
@@ -235,25 +237,27 @@ final class Settings_Page {
 			$field_args
 		);
 
-		// API Secret.
-		$field_id   = 'api_secret';
-		$field_args = array(
-			'option_key'    => $field_id,
-			'help_text'     => __( 'Your API secret is your secret code to <a href="https://docs.parse.ly/the-parsely-api/">access our API</a>. It can be found at <code>dash.parsely.com/<var>yoursitedomain</var>/settings/api</code> (replace <var>yoursitedomain</var> with your domain name, e.g. <samp>mydomain.com</samp>).<br />If you haven\'t purchased access to the API and would like to do so, email your account manager or <a href="mailto:support@parsely.com">support@parsely.com</a>.', 'wp-parsely' ),
-			'label_for'     => $field_id,
-			'optional_args' => array(
-				'type'                => 'password',
-				'is_obfuscated_value' => true,
-			),
-		);
-		add_settings_field(
-			$field_id,
-			__( 'API Secret', 'wp-parsely' ),
-			array( $this, 'print_text_tag' ),
-			Parsely::MENU_SLUG,
-			$section_key,
-			$field_args
-		);
+		if ( ! $are_credentials_managed ) {
+			// API Secret.
+			$field_id   = 'api_secret';
+			$field_args = array(
+				'option_key'    => $field_id,
+				'help_text'     => __( 'Your API secret is your secret code to <a href="https://docs.parse.ly/the-parsely-api/">access our API</a>. It can be found at <code>dash.parsely.com/<var>yoursitedomain</var>/settings/api</code> (replace <var>yoursitedomain</var> with your domain name, e.g. <samp>mydomain.com</samp>).<br />If you haven\'t purchased access to the API and would like to do so, email your account manager or <a href="mailto:support@parsely.com">support@parsely.com</a>.', 'wp-parsely' ),
+				'label_for'     => $field_id,
+				'optional_args' => array(
+					'type'                => 'password',
+					'is_obfuscated_value' => true,
+				),
+			);
+			add_settings_field(
+				$field_id,
+				__( 'API Secret', 'wp-parsely' ),
+				array( $this, 'print_text_tag' ),
+				Parsely::MENU_SLUG,
+				$section_key,
+				$field_args
+			);
+		}
 
 		// Metadata Format.
 		$field_id   = 'meta_type';
@@ -517,7 +521,8 @@ final class Settings_Page {
 	 * @since 3.2.0
 	 */
 	private function initialize_advanced_section(): void {
-		$section_key = 'advanced-section';
+		$are_credentials_managed = Parsely::are_credentials_managed();
+		$section_key             = 'advanced-section';
 
 		add_settings_section(
 			$section_key,
@@ -526,25 +531,27 @@ final class Settings_Page {
 			Parsely::MENU_SLUG
 		);
 
-		// Metadata Secret.
-		$field_id   = 'metadata_secret';
-		$field_args = array(
-			'option_key'    => $field_id,
-			'help_text'     => __( 'Your metadata secret is given to you by Parse.ly support. DO NOT enter anything here unless given to you by Parse.ly support!', 'wp-parsely' ),
-			'label_for'     => $field_id,
-			'optional_args' => array(
-				'type'                => 'password',
-				'is_obfuscated_value' => true,
-			),
-		);
-		add_settings_field(
-			$field_id,
-			__( 'Metadata Secret', 'wp-parsely' ),
-			array( $this, 'print_text_tag' ),
-			Parsely::MENU_SLUG,
-			$section_key,
-			$field_args
-		);
+		if ( ! $are_credentials_managed ) {
+			// Metadata Secret.
+			$field_id   = 'metadata_secret';
+			$field_args = array(
+				'option_key'    => $field_id,
+				'help_text'     => __( 'Your metadata secret is given to you by Parse.ly support. DO NOT enter anything here unless given to you by Parse.ly support!', 'wp-parsely' ),
+				'label_for'     => $field_id,
+				'optional_args' => array(
+					'type'                => 'password',
+					'is_obfuscated_value' => true,
+				),
+			);
+			add_settings_field(
+				$field_id,
+				__( 'Metadata Secret', 'wp-parsely' ),
+				array( $this, 'print_text_tag' ),
+				Parsely::MENU_SLUG,
+				$section_key,
+				$field_args
+			);
+		}
 
 		// Disable autotrack.
 		add_settings_field(
@@ -674,7 +681,7 @@ final class Settings_Page {
 		$name                = Parsely::OPTIONS_KEY . "[$id]";
 		$is_obfuscated_value = $optional_args['is_obfuscated_value'] ?? false;
 		$value               = $is_obfuscated_value ? $this->get_obfuscated_value( $value ) : esc_attr( $value );
-		$accepted_args       = array( 'placeholder', 'required' );
+		$accepted_args       = array( 'placeholder', 'required', 'disabled' );
 		$type                = $optional_args['type'] ?? 'text';
 
 		echo sprintf( "<input type='%s' name='%s' id='%s' value='%s'", esc_attr( $type ), esc_attr( $name ), esc_attr( $id ), esc_attr( $value ) );
@@ -684,8 +691,12 @@ final class Settings_Page {
 		}
 
 		foreach ( $optional_args as $key => $val ) {
-			if ( \in_array( $key, $accepted_args, true ) ) {
-				echo ' ' . esc_attr( $key ) . '="' . esc_attr( (string) $val ) . '"';
+			if ( \in_array( $key, $accepted_args, true ) && false !== $val ) {
+				// Support attributes without values.
+				echo ' ' . esc_attr( $key );
+				if ( true !== $val ) {
+					echo '="' . esc_attr( $val ) . '"';
+				}
 			}
 		}
 		echo ' />';
@@ -930,35 +941,41 @@ final class Settings_Page {
 	 * @return ParselySettingOptions Validated inputs.
 	 */
 	private function validate_basic_section( $input ) {
-		$options = $this->parsely->get_options();
+		$are_credentials_managed = Parsely::are_credentials_managed();
+		$options                 = $this->parsely->get_options();
 
-		if ( '' === $input['apikey'] ) {
-			add_settings_error(
-				Parsely::OPTIONS_KEY,
-				'apikey',
-				__( 'Please specify the Site ID', 'wp-parsely' )
-			);
+		if ( $are_credentials_managed ) {
+			$input['apikey']     = '';
+			$input['api_secret'] = '';
 		} else {
-			$site_id = $this->sanitize_site_id( $input['apikey'] );
-			if ( false === $this->validate_site_id( $site_id ) ) {
+			if ( '' === $input['apikey'] ) {
 				add_settings_error(
 					Parsely::OPTIONS_KEY,
 					'apikey',
-					__( 'The Site ID is incorrect. It should look like "example.com".', 'wp-parsely' )
+					__( 'Please specify the Site ID', 'wp-parsely' )
 				);
 			} else {
-				$input['apikey'] = $site_id;
+				$site_id = $this->sanitize_site_id( $input['apikey'] );
+				if ( false === $this->validate_site_id( $site_id ) ) {
+					add_settings_error(
+						Parsely::OPTIONS_KEY,
+						'apikey',
+						__( 'The Site ID is incorrect. It should look like "example.com".', 'wp-parsely' )
+					);
+				} else {
+					$input['apikey'] = $site_id;
+				}
 			}
-		}
 
-		$input['api_secret'] = $this->get_unobfuscated_value( $input['api_secret'], $this->parsely->get_api_secret() );
-		$api_secret_length   = strlen( $input['api_secret'] );
-		if ( $api_secret_length > 0 && $api_secret_length <= 30 ) {
-			add_settings_error(
-				Parsely::OPTIONS_KEY,
-				'api_secret',
-				__( 'The API Secret is incorrect. Please contact Parse.ly support!', 'wp-parsely' )
-			);
+			$input['api_secret'] = $this->get_unobfuscated_value( $input['api_secret'], $this->parsely->get_api_secret() );
+			$api_secret_length   = strlen( $input['api_secret'] );
+			if ( $api_secret_length > 0 && $api_secret_length <= 30 ) {
+				add_settings_error(
+					Parsely::OPTIONS_KEY,
+					'api_secret',
+					__( 'The API Secret is incorrect. Please contact Parse.ly support!', 'wp-parsely' )
+				);
+			}
 		}
 
 		if ( ! isset( $input['meta_type'] ) ) {
@@ -1130,23 +1147,28 @@ final class Settings_Page {
 	 * @return ParselySettingOptions Validated inputs.
 	 */
 	private function validate_advanced_section( $input ) {
-		$options = $this->parsely->get_options();
+		$are_credentials_managed = Parsely::are_credentials_managed();
+		$options                 = $this->parsely->get_options();
 
-		$input['metadata_secret'] = $this->get_unobfuscated_value( $input['metadata_secret'], $this->parsely->get_options()['metadata_secret'] );
-		if ( '' !== $input['metadata_secret'] ) {
-			if ( strlen( $input['metadata_secret'] ) !== 10 ) {
-				add_settings_error(
-					Parsely::OPTIONS_KEY,
-					'metadata_secret',
-					__( 'The Metadata Secret is incorrect. Please contact Parse.ly support!', 'wp-parsely' )
-				);
-			} elseif (
-				isset( $input['parsely_wipe_metadata_cache'] ) && 'true' === $input['parsely_wipe_metadata_cache'] // @phpstan-ignore-line
-			) {
-				delete_post_meta_by_key( 'parsely_metadata_last_updated' );
+		if ( $are_credentials_managed ) {
+			$input['metadata_secret'] = '';
+		} else {
+			$input['metadata_secret'] = $this->get_unobfuscated_value( $input['metadata_secret'], $this->parsely->get_options()['metadata_secret'] );
+			if ( '' !== $input['metadata_secret'] ) {
+				if ( strlen( $input['metadata_secret'] ) !== 10 ) {
+					add_settings_error(
+						Parsely::OPTIONS_KEY,
+						'metadata_secret',
+						__( 'The Metadata Secret is incorrect. Please contact Parse.ly support!', 'wp-parsely' )
+					);
+				} elseif (
+					isset( $input['parsely_wipe_metadata_cache'] ) && 'true' === $input['parsely_wipe_metadata_cache'] // @phpstan-ignore-line
+				) {
+					delete_post_meta_by_key( 'parsely_metadata_last_updated' );
 
-				wp_schedule_event( time() + 100, 'everytenminutes', 'parsely_bulk_metas_update' );
-				$input['parsely_wipe_metadata_cache'] = false;
+					wp_schedule_event( time() + 100, 'everytenminutes', 'parsely_bulk_metas_update' );
+					$input['parsely_wipe_metadata_cache'] = false;
+				}
 			}
 		}
